+++ SPOT/DataAccess/Sources/RunningDataAccess/LocationClient.swift
@@ -6,14 +6,50 @@
 //
 
 import CoreLocation
-import Dependencies
-import DependenciesWrapper
+import Running
+
+//protocol LocationClient {
+//    func delegate(completion: @Sendable () async -> AsyncStream<RunningLocationManager.Action>)
+//    func authorizationStatus(completion: @Sendable () async -> CLAuthorizationStatus)
+//    func location(completion: @Sendable () async -> Location)
+//    func requestLocation(completion: @Sendable () async -> Void)
+//    func requestWhenInUseAuthorization(completion: @Sendable () async -> Void)
+//    func startUpdatingLocation(completion: @Sendable () async -> Void)
+//}
+
+@MainActor
+public final class MainActorIsolated<Value>: Sendable {
+    public lazy var value: Value = initialValue()
+    private let initialValue: @MainActor () -> Value
+    nonisolated public init(initialValue: @MainActor @escaping () -> Value) {
+        self.initialValue = initialValue
+    }
+}
 
 public struct LcationAuthorizationStatus: Equatable {
     let status: CLAuthorizationStatus
 }
 
-public struct LocationClient {
+public final class RunningLocationManager: LocationUsecaseProtocol {
+    private let manager: MainActorIsolated<CLLocationManager>
+    
+    public init(manager: MainActorIsolated<CLLocationManager> = MainActorIsolated(initialValue: { CLLocationManager() })) {
+        self.manager = manager
+    }
+    
+    public func start() {
+        
+        manager.value.delegate
+    }
+    
+    public func pause() {
+        
+    }
+    
+    public func stop() {
+        
+    }
+    
     public enum Action {
         case didChangeAuthorization(CLAuthorizationStatus)
         case didUpdateLocations([Location])
@@ -29,34 +65,41 @@ public struct LocationClient {
             self.error = error as NSError
         }
     }
+}
+
+private final class RunningLocationManagerDelegate: NSObject, CLLocationManagerDelegate {
+    let continuation: AsyncStream<RunningLocationManager.Action>.Continuation
     
-    public var delegate: @Sendable () async -> AsyncStream<LocationClient.Action>
-    public var authorizationStatus: @Sendable () async -> CLAuthorizationStatus
-    public var location: @Sendable () async -> Location?
-    public var requestLocation: @Sendable () async -> Void
-    public var requestWhenInUseAuthorization: @Sendable () async -> Void
-    public var startUpdatingLocation: @Sendable () async -> Void
-}
-
-extension LocationClient {
-    public static let test = Self(
-        delegate: unimplemented("\(Self.self).delegate"),
-        authorizationStatus: unimplemented("\(Self.self).authorizationStatus"),
-        location: unimplemented("\(Self.self).location"),
-        requestLocation: unimplemented("\(Self.self).requestLocation"),
-        requestWhenInUseAuthorization: unimplemented("\(Self.self).requestWhenInUseAuthorization"),
-        startUpdatingLocation: unimplemented("\(Self.self).startUpdatingLocation")
-    )
-}
-
-extension DependencyValues {
-    public var locationClient: LocationClient {
-        get { self[LocationClient.self] }
-        set { self[LocationClient.self] = newValue }
+    public init(
+        continuation: AsyncStream<RunningLocationManager.Action>.Continuation
+    ) {
+        self.continuation = continuation
+    }
+    
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        continuation.yield(
+            .didUpdateLocations(
+                locations.map {
+                    Location(coordinate: .init(latitude: $0.coordinate.latitude, longitude: $0.coordinate.longitude))
+                }
+            )
+        )
+    }
+    
+    func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
+        continuation.yield(.didChangeAuthorization(manager.authorizationStatus))
+    }
+    
+    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
+        continuation.yield(.didFailWithError(RunningLocationManager.Error(error)))
+    }
+    
+    func locationManagerDidPauseLocationUpdates(_ manager: CLLocationManager) {
+        continuation.yield(.didPauseLocationUpdates)
+    }
+    
+    func locationManagerDidResumeLocationUpdates(_ manager: CLLocationManager) {
+        continuation.yield(.didResumeLocationUpdates)
     }
 }
 
-extension LocationClient: DependencyKey {
-    public static var liveValue: Self { .live }
-    public static var testValue: Self { .test }
-}
