+++ SPOT/DataAccess/Sources/RunningDataAccess/LocationClientLive.swift
@@ -42,49 +42,3 @@ extension LocationClient {
         )
     }
 }
-
-private final class LocationClientDelegate: NSObject, CLLocationManagerDelegate {
-    let continuation: AsyncStream<LocationClient.Action>.Continuation
-    
-    public init(
-        continuation: AsyncStream<LocationClient.Action>.Continuation
-    ) {
-        self.continuation = continuation
-    }
-    
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        continuation.yield(
-            .didUpdateLocations(
-                locations.map {
-                    Location(coordinate: .init(latitude: $0.coordinate.latitude, longitude: $0.coordinate.longitude))
-                }
-            )
-        )
-    }
-    
-    func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
-        continuation.yield(.didChangeAuthorization(manager.authorizationStatus))
-    }
-    
-    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
-        continuation.yield(.didFailWithError(LocationClient.Error(error)))
-    }
-    
-    func locationManagerDidPauseLocationUpdates(_ manager: CLLocationManager) {
-        continuation.yield(.didPauseLocationUpdates)
-    }
-    
-    func locationManagerDidResumeLocationUpdates(_ manager: CLLocationManager) {
-        continuation.yield(.didResumeLocationUpdates)
-    }
-}
-
-@MainActor
-public final class MainActorIsolated<Value>: Sendable {
-    public lazy var value: Value = initialValue()
-    private let initialValue: @MainActor () -> Value
-    nonisolated public init(initialValue: @MainActor @escaping () -> Value) {
-        self.initialValue = initialValue
-    }
-}
-
